+++ src/Adapter/FloatCallbackAdapter.php
@@ -1,0 +1,47 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Cacing69\Cquery\Adapter;
+
+use Cacing69\Cquery\CallbackAdapter;
+
+class FloatCallbackAdapter extends CallbackAdapter
+{
+    protected static $signature = '/^\s*float\(\s*(.*?)\s*\)\s*$/is';
+
+    public static function getSignature()
+    {
+        return self::$signature;
+    }
+
+    public function __construct(string $raw)
+    {
+        $this->raw = $raw;
+
+        $this->callback = function (string $value) {
+            return floatval($value);
+        };
+
+        // check if function is nested
+        if (preg_match('/^\s?int\(\s?([a-z0-9_]*\(.+?\))\s?\)$/', $raw)) {
+            preg_match('/^\s?int\(\s?([a-z0-9_]*\(.+?\))\s?\)$/', $raw, $extract);
+
+            $extractChild = $this->extractChild($extract[1]);
+            $_childCallback = $extractChild->getAdapter()->getCallback();
+
+            if($_childCallback) {
+                $this->callback = function (string $value) use ($_childCallback) {
+                    return floatval((string) $_childCallback($value));
+                };
+            }
+        } else {
+            preg_match(self::$signature, $raw, $node);
+
+            $this->node = $node[1];
+
+            $this->callMethod = "extract";
+            $this->callMethodParameter = ["_text"];
+        }
+    }
+}