+++ src/RegisterAdapter.php
@@ -5,10 +5,14 @@ namespace Cacing69\Cquery;
 use Cacing69\Cquery\Adapter\AttributeCallbackAdapter;
 use Cacing69\Cquery\Adapter\DefaultCallbackAdapter;
 use Cacing69\Cquery\Adapter\AppendNodeCallbackAdapter;
+use Cacing69\Cquery\Adapter\FloatCallbackAdapter;
+use Cacing69\Cquery\Adapter\IntegerCallbackAdapter;
 use Cacing69\Cquery\Adapter\LengthCallbackAdapter;
 use Cacing69\Cquery\Adapter\LowerCallbackAdapter;
 use Cacing69\Cquery\Adapter\ReplaceCallbackAdapter;
 use Cacing69\Cquery\Adapter\ReverseCallbackAdapter;
+use Cacing69\Cquery\Adapter\SingleQuotesCallbackAdapter;
+use Cacing69\Cquery\Adapter\StringCallbackAdapter;
 use Cacing69\Cquery\Adapter\UpperCallbackAdapter;
 
 /**
@@ -32,6 +36,10 @@ class RegisterAdapter
             LowerCallbackAdapter::class,
             LowerCallbackAdapter::class,
             ReplaceCallbackAdapter::class,
+            SingleQuotesCallbackAdapter::class,
+            IntegerCallbackAdapter::class,
+            StringCallbackAdapter::class,
+            FloatCallbackAdapter::class,
 
             /**
              *  Make sure that DefaultCallbackAdapter is always at the bottom