+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Release Notes
 
 ## [v1.1.0](https://github.com/cacing69/cquery/compare/v1.0.0...v1.0.1) (2023-0x-xx)
-- Add method onReady(Closure $closure) for submit or go to another page when firstPage loaded
+- Add method onContentLoaded(Closure $closure) for submit or go to another page when content loaded
+- Add method eachItem(Closure $closure) to manipulate each item
+- Add method onObtainedResults(Closure $closure) to manipulate result array with closure callback
+- Add single quotes adapter
+- Add String adapter
+- Add Integer adapter
+- Add Float adapter
 ## [v1.0.0](https://github.com/cacing69/cquery/releases/tag/v1.0.0) (2023-09-05)
 
 - First release